+++ src/main/java/com/example/demo/repository/EventRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventRepo extends JpaRepository<Event, Integer> {
+public interface EventRepo extends JpaRepository<Event, Long> {
 
 }
